+++ packages/post-processors/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 1.3.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/shared@0.5.1
+  - @volontariapp/database@3.2.2
+  - @volontariapp/messaging@2.3.2
+  - @volontariapp/outbox@0.9.5
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.3.1
+  - @volontariapp/database@3.2.1
+  - @volontariapp/outbox@0.9.4
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.3.0
+  - @volontariapp/database@3.2.0
+  - @volontariapp/outbox@0.9.3
+
 ## 1.3.2
 
 ### Patch Changes